Colour method is now in view model, and colour styles folder
