+++ fireapp/lib/style/colors.dart
@@ -1,0 +1,12 @@
+import 'dart:ui';
+
+const List<Color> calendarEventColourPalette = [
+  Color(0xFFB9DEEC), // Light Blue
+  Color(0xFFCBCBEA), // Lavender
+  Color(0xFFFDD2D7), // Pink
+  Color(0xFFFCFCDC), // Light Yellow
+  Color(0xFFEFEFEF), // Light Gray
+  Color(0xFFF6CEBC), // Light Orange
+  Color(0xFFCFDEE3),
+  Color(0xFFCBE4F1),
+];
